[3019] Move the widget id provider to sirius-components-forms
Bug: https://github.com/eclipse-sirius/sirius-web/issues/3019
