+++ packages/forms/backend/sirius-components-forms/src/main/java/org/eclipse/sirius/components/forms/WidgetIdProvider.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019, 2023 Obeo.
+ * Copyright (c) 2019, 2024 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -10,7 +10,7 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-package org.eclipse.sirius.components.compatibility.forms;
+package org.eclipse.sirius.components.forms;
 
 import java.util.UUID;
 import java.util.function.Function;
